+++ react.events/src/App.js
@@ -4,7 +4,7 @@ import './App.css';
 import {BrowserRouter} from "react-router-dom";
 import {Routes, Route} from "react-router-dom";
 import HomePage from "./containers/Pages/HomePage/HomePage";
-import Events from "./containers/Pages/EventsPage/Events";
+import AppView from "./containers/Pages/EventsPage/AppView";
 import Favorite from "./containers/Pages/FavoritePage/Favorite";
 import About from "./containers/Pages/AboutPage/About";
 import Login from "./containers/Pages/LoginPage/Login";
@@ -19,7 +19,7 @@ function App() {
                 <HeaderContainer/>
                 <Routes>
                     <Route path='/' element={<HomePage/>}/>
-                    <Route path='/events' element={<Events/>}/>
+                    <Route path='/events' element={<AppView/>}/>
                     <Route path='/favorite' element={<Favorite/>}/>
                     <Route path='/about' element={<About/>}/>
                     <Route path='/login' element={<Login/>}/>
